refactor(multiadmin): move to services/ and merge admin/server
- Move multiadmin package to go/services/multiadmin/
- Merge go/admin/server/ files (backup.go, server.go, server_test.go) into the same package
- Add multiadmin-isolation depguard rule to prevent other packages from importing the service code directly
