+++ go/services/multiadmin/discovery.go
